pyproject.toml
[project]
name = "spotify-mcp"
version = "0.1.0"
description = "Production Spotify MCP Server with OAuth 2.0 support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.5.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "spotipy>=2.22.1",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.0.291",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
